Clean up browser comparison chart
